requirements.txt
pandas==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
yfinance==0.2.28
apscheduler==3.10.4
flask==2.2.5
plotly==5.16.1
dash==2.13.0
faker==19.3.1